WIP Rename testdefinition to evaluations
(cherry picked from commit 389d2562fd9494385a3c5f9c59b7ab1259948b67)
