+++ packages/frontend/editor-ui/src/components/Evaluations/EditDefinition/sections/RunsSection.vue
@@ -1,7 +1,7 @@
 <script setup lang="ts">
 import type { TestRunRecord } from '@/api/testDefinition.ee';
-import MetricsChart from '@/components/TestDefinition/ListRuns/MetricsChart.vue';
-import TestRunsTable from '@/components/TestDefinition/ListRuns/TestRunsTable.vue';
+import MetricsChart from '@/components/Evaluations/ListRuns/MetricsChart.vue';
+import TestRunsTable from '@/components/Evaluations/ListRuns/TestRunsTable.vue';
 import { useI18n } from '@/composables/useI18n';
 import { VIEWS } from '@/constants';
 import { convertToDisplayDate } from '@/utils/formatters/dateFormatter';
