+++ packages/frontend/editor-ui/src/components/Evaluations/shared/TableStatusCell.vue
@@ -1,7 +1,7 @@
 <script setup lang="ts" generic="T">
 import N8nTooltip from '@n8n/design-system/components/N8nTooltip';
 import type { BaseTextKey } from '@/plugins/i18n';
-import type { TestTableColumn } from '@/components/TestDefinition/shared/TestTableBase.vue';
+import type { TestTableColumn } from '@/components/Evaluations/shared/TestTableBase.vue';
 import { useI18n } from '@/composables/useI18n';
 import { useRouter } from 'vue-router';
