+++ packages/frontend/editor-ui/src/views/Evaluations/TestDefinitionListView.vue
@@ -1,6 +1,6 @@
 <script setup lang="ts">
-import EmptyState from '@/components/TestDefinition/ListDefinition/EmptyState.vue';
-import TestItem from '@/components/TestDefinition/ListDefinition/TestItem.vue';
+import EmptyState from '@/components/Evaluations/ListDefinition/EmptyState.vue';
+import TestItem from '@/components/Evaluations/ListDefinition/TestItem.vue';
 import { useI18n } from '@/composables/useI18n';
 import { useMessage } from '@/composables/useMessage';
 import { useToast } from '@/composables/useToast';
